requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
requests==2.21.0
todoist-python==7.0.18
urllib3==1.24.1
